+++ src/main/java/ru/yandex/practicum/sleeptracker/functions/TotalSessionsFunction.java
@@ -9,7 +9,7 @@ import java.util.function.Function;
 public class TotalSessionsFunction implements Function<List<SleepingSession>, SleepAnalysisResult> {
     @Override
     public SleepAnalysisResult apply(List<SleepingSession> listOfSessions) {
-        int count = (listOfSessions == null) ? 0 : listOfSessions.size();
+        long count = (listOfSessions == null) ? 0 : listOfSessions.size();
         return new SleepAnalysisResult("Количество сессий сна за представленный период", count);
     }
 }
